requirements.txt
asgiref==3.6.0
Django==4.1.7
django-location-field==2.7.0
django-phonenumber-field==7.0.2
geographiclib==2.0
geopy==2.3.0
phonenumbers==8.13.7
Pillow==9.4.0
psycopg2-binary==2.9.5
six==1.16.0
sqlparse==0.4.3
